Fix go tests in /lib/go/test/tests.
